embedding-service/requirements.txt
fastapi
uvicorn
qdrant-client==1.9.0
boto3
google-generativeai 